Add resources for game page
This page has been almost completley converted to a client side
implementation. All of the styling and waveforms/canvas based objects
work as expected, and all events/dynamic content has been assosiated
with a Vue instance :D And It already has localisation baked in, really
as a consiquence of how I coded the server/client side originally and
how the Vue migration naturally fell out (To fix for Canada I just need
to add loc buttons on the game page)

However several components are burned in. For example the system
parameters and the waveform source files. All of the burned in content
is certainly ready to be made dynamic, and for sure the hardest bits of
the migration are done in this commit :D

This also does not handle level progression. The level number is also
burned in on the game page, but it has been tested that the game page
works for both level 1 (with the tutorial banner) and for higher levels
(with the quote banner).

Handling level progression will be done through cookies, and while I
haven't really got much experience with cookie manipulation, I have the
`getCookie` method already from w3schools and just pulled in a
`setCookie` method. So again I belive the hardest migration problems
have been solved :D