Rm unspec submit meth with spec state change meths
The `submit` method was unspecific, taking someone with any number of
lives with on any level to either the next level, the complete or lost
page.

This unspecific method has been replaces with three focused methods that
handle specific state changes; next level, won, lost.

The won and lost methods are essentially complete. I cannot see any
extra logic that needs to go into these methods right now, unless I
decided to reset the lives and levels in these methods rather than on
the page they point at. However the pages they point at are currently
place holders for documents that don't exist in this commit.

The next level method does take you to the game page that exsits.
However the cookie recorded state or lives and level are not updated
before the next level method is triggered, so you are simply returned to
the original `(level, lives) = (1, 3)` state.