Set level+lives cookies on game state change
This addition enables progression throught the game and ensures a reset
state if the game is won or lost. If the game is quit (for example by
closing the tab) and the user returns directley to the `/game.html` page
then they will start from the level they left at with the number of
lives they had (assuming no cookies were tampered with).

This later outcome was not targeted, however I think that it is a
reasonable outcome to have.