Rm countdown method add won+lost countdown methods
The beginCountdown method was supposed to handle both won and lost state
changes. However it depended on submit previously and in refactoring as
three specific methods a bug was introduced where the beginCountdown
method would only transition to a lost state even if the target state
was a won.

Two new methods targeting a won and lost state noe replace this single
method.

The level and lives cookies are also reset at the begining of the
countdown methods, rather than at the point of transition. This was
motivated by an edge case where a user who has run out of lives could
reload the page before the countdown has finished to be brought back to
the lives and level state they were in before they lost their final
life. They would be given a new set of waveforms so it doesn't
nessesarily matter, but a reset on win or lose is the intended result so
ensuring consitencey is important here.