Up package.json right info; Add production webpack
Added to the package.json are a couple of user script string to both
spin up a webpack-dev-server and also build a static production webpack.
They don't really add any extra convinience, but usefull to remember the
right commands to do those task.

The package.json was also poluted with info from spectrum, corrected in
this commit.