Rm cookie setter path attribute
The path attruibute is redundant

- Cookies cannot be set on pages. To the best of my knowlegde if a
  specific page is passed as the path attribute then the cookie is set
  on the parent folder.
- If not specified then the path attribute is set to the parent folder
  of the current page.
- All pages in the site sit in the root of the site. By the first and
  second bullet, not explicitly setting the path attribute will achive
  the same as explicitly setting the path attribute.

There is a final issue this resolves. When the site is not deployed at
the root of the domain, the `path=/game.html` or `path=/` specifications
would not restrict the cookie to the deployed path. In theory they could
be overridden by other sites on the domain.