Change all paths to relative paths
Deplying on a subdirectory doesn't work. While absolutle paths are
better for SEO, this is not a concern at the moment and a simple fix
to the issue.

Also note that several paths on root pages were prefixed with `../`.
They did appear to work when running a dev server. By this I belive that
any `../` when you are already at the root of the domain resolves to
`./` which is why the issue passed silently.