Add build
For this major version series of `black-hole-hunter`, separating build
branches and release etc might be the intelegent thing, but it creates
much more hassel for what is a very simple build output; one directory
of 4 js files.

So for this major version series the build will be included along with
the main development branch for convinience.