Make game content fill window
Mostly this was correctly applying flexbox styles. One non-trivial issue
was make the `#game` div take the correct height (which in turn
determines the heights aquired by its decendant wavesurferes).

The issue was that the height of the `#game` `div` should resize based
on the other fixed hight *rows*. The carousel has to have its height
dynamically resized. The issue is that the `#game` `div` does not
respect the resized height, which either caused `#game` to overflow the
height of the viewport, or more bizzarly, just the fullsize wavesurfer
to be generated taller than its container, which almost completly
occludes the player controls.

This is a bug in the browser engine, as is reported on SE and else
where. The solution is to force a redraw, and this is done in
`game.coffeeL453`. For some reason the `0` in `show` is nessesary. By
the doc's this causes a visible animation of 0s.