Move globnav margins from children to decendants
With the margins (top/bottom) on the children of the `global-nav`
flexbox, a vertical scroll appears even though there is nothing to
scroll.

Moving the margins from the children to deeper decendants fixes this
behaviour. Its not clear why this happens and I would speculate this is
another bug in the rendering engine.