Freeze Vue version to restore functionality
Black Hole Hunter stoped working, noticed on 2022-03-02. It appears that recent 3.X.X versions of Vue no longer have the `vue.js` bundle on unpkg. The version has been frozen to the last 2.X.X release which has a `vue.js` bundle (although now we pount at the `vue.min.js` bundle) and should function. Will freeze to an eariler version in a follow up commit if this is too recent a version.